feat!: 2.0.0
This release adds support for the PyPI RSS feeds and the current versions of the PyPI JSON API
and the Simple Repository API. It also removes a lot of deprecated/unused attributes.

This commit, apart from the changelog, also includes some tiny updates to the documentation
and a new `metadata_url` attribute.